requirements.txt
fastapi>=0.104.1
pydantic<2.0.0
uvicorn[standard]>=0.23.2
python-multipart>=0.0.6
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0.76
numpy>=1.24.3
python-dotenv>=1.0.0
a2wsgi>=1.7.0

easyocr>=1.7.1
torch>=2.1.0
torchvision>=0.16.0
